requirements.txt
psutil
yara-python
pyvelociraptor
pyzipper
Jinja2
oletools
requests
python-dotenv
redis
